Code cleaning, smoothed camera movement
Cleaned up the code:
- removed obsolete code
- renamed scenes and nodes
- organized scenes and routines

Improvements on Camera Movement:
- Camera movement can be enabled and disabled changing a single flag
- Camera moves more smoothly
- Camera reacts to a change in window size - reactive are changes
automatically
- Camera moves to the active character at the beginning of every turn
- Menu is closed until active character is in focus
- Camera can be moved by the player and camera speed is adjusted
depending on how much the mouse is close to the borders of the window
- Manual movement of the camera is activated only when shift button is
pressed

Bug Report:
- Solved
- Bug that caused the menu to react if it appears right where selectable
area was selected during previous turn.
- New
- Even if the game begins with change_state("Action_Select") this
doesn't happen if the camera is disabled. Probable Cause: yield() is
used to wait for the camera to be moved, if camera is disabled the
signal is probably emitted before yield is called. This is not an issue
in successive turns.